Move ruleset and layouts into top level
- ruleset uses Ball objects which is problematic for fully-connected
  subpackage sharing with GameType
