--- pooltool/game/ruleset/snooker/__init__.py
@@ -1,5 +1,0 @@
-from pooltool.game.ruleset.snooker.core import _Snooker
-
-__all__ = [
-    "_Snooker",
-]
