+++ pooltool/ruleset/snooker/__init__.py
@@ -1,0 +1,5 @@
+from pooltool.ruleset.snooker.core import _Snooker
+
+__all__ = [
+    "_Snooker",
+]
