+++ pooltool/ruleset/sum_to_three.py
@@ -4,9 +4,9 @@ from __future__ import annotations
 
 from typing import Counter
 
-from pooltool.events.datatypes import Event, EventType
-from pooltool.events.filter import by_ball, by_time, by_type, filter_events
-from pooltool.game.ruleset.datatypes import (
+from pooltool.events.datatypes import EventType
+from pooltool.events.filter import by_ball, by_type, filter_events, filter_type
+from pooltool.ruleset.datatypes import (
     BallInHandOptions,
     Player,
     Ruleset,
@@ -16,44 +16,27 @@ from pooltool.game.ruleset.datatypes import (
 from pooltool.system.datatypes import System
 
 
-def _other(cue: str, event: Event) -> str:
-    for id in event.ids:
-        if id != cue:
-            return id
-
-    raise Exception()
-
-
-def is_turn_over(shot: System, constraints: ShotConstraints) -> bool:
-    assert constraints.cueable is not None
-    cue = constraints.cueable[0]
-
-    # Find when the second ball is first hit by the cue-ball
-
-    ball_hits = filter_events(
-        shot.events,
-        by_type(EventType.BALL_BALL),
-        by_ball(cue),
-    )
-
-    hits = set()
-    for event in ball_hits:
-        hits.add(_other(cue, event))
-        if len(hits) == 2:
-            break
-    else:
+def is_turn_over(shot: System) -> bool:
+    # See whether cue contacted object ball
+    ball_hits = filter_type(shot.events, EventType.BALL_BALL)
+    if not len(ball_hits):
         return True
 
-    # Now calculate all cue-ball cushion hits before that event
-
-    cushion_hits = filter_events(
+    # Count rails that cue ball hits
+    cue_cushion_hits = filter_events(
         shot.events,
         by_type(EventType.BALL_LINEAR_CUSHION),
-        by_ball(cue),
-        by_time(event.time, after=False),
+        by_ball("cue"),
     )
 
-    return len(cushion_hits) < 3
+    # Count rails that object ball hits
+    object_cushion_hits = filter_events(
+        shot.events,
+        by_type(EventType.BALL_LINEAR_CUSHION),
+        by_ball("object"),
+    )
+
+    return len(cue_cushion_hits) + len(object_cushion_hits) != 3
 
 
 def is_game_over(
@@ -65,23 +48,13 @@ def is_game_over(
     return score[active.name] == win_condition
 
 
-def next_cue(current_cue: str, num_players: int) -> str:
-    assert current_cue in ("white", "yellow", "red")
-    assert num_players in (2, 3)
-
-    if num_players == 3:
-        return {"white": "yellow", "yellow": "red", "red": "white"}[current_cue]
-
-    return "white" if current_cue == "yellow" else "yellow"
-
-
-class _ThreeCushion(Ruleset):
+class _SumToThree(Ruleset):
     def __init__(self, *args, win_condition: int = 10, **kwargs):
         self.win_condition = win_condition
         Ruleset.__init__(self, *args, **kwargs)
 
     def build_shot_info(self, shot: System) -> ShotInfo:
-        turn_over = is_turn_over(shot, self.shot_constraints)
+        turn_over = is_turn_over(shot)
         score = self.get_score(self.score, turn_over)
         game_over = is_game_over(
             score,
@@ -104,24 +77,13 @@ class _ThreeCushion(Ruleset):
         return ShotConstraints(
             ball_in_hand=BallInHandOptions.NONE,
             movable=[],
-            cueable=["white"],
-            hittable=tuple(),
+            cueable=["cue"],
+            hittable=("object",),
             call_shot=False,
         )
 
     def next_shot_constraints(self, shot: System) -> ShotConstraints:
-        assert (cueable := self.shot_constraints.cueable) is not None
-
-        if self.shot_info.turn_over:
-            cueable = [next_cue(cueable[0], len(self.players))]
-
-        return ShotConstraints(
-            ball_in_hand=BallInHandOptions.NONE,
-            movable=[],
-            cueable=cueable,
-            hittable=tuple(),
-            call_shot=False,
-        )
+        return self.shot_constraints
 
     def get_score(self, score: Counter, turn_over: bool) -> Counter:
         if turn_over:
@@ -140,5 +102,5 @@ class _ThreeCushion(Ruleset):
         if self.shot_info.turn_over:
             self.log.add_msg(f"{self.last_player.name} is up!", sentiment="good")
 
-    def copy(self) -> _ThreeCushion:
+    def copy(self) -> _SumToThree:
         raise NotImplementedError("ThreeCushion copy needs to be implemented")
