+++ pooltool/ruleset/three_cushion.py
@@ -4,9 +4,9 @@ from __future__ import annotations
 
 from typing import Counter
 
-from pooltool.events.datatypes import EventType
-from pooltool.events.filter import by_ball, by_type, filter_events, filter_type
-from pooltool.game.ruleset.datatypes import (
+from pooltool.events.datatypes import Event, EventType
+from pooltool.events.filter import by_ball, by_time, by_type, filter_events
+from pooltool.ruleset.datatypes import (
     BallInHandOptions,
     Player,
     Ruleset,
@@ -16,27 +16,44 @@ from pooltool.game.ruleset.datatypes import (
 from pooltool.system.datatypes import System
 
 
-def is_turn_over(shot: System) -> bool:
-    # See whether cue contacted object ball
-    ball_hits = filter_type(shot.events, EventType.BALL_BALL)
-    if not len(ball_hits):
+def _other(cue: str, event: Event) -> str:
+    for id in event.ids:
+        if id != cue:
+            return id
+
+    raise Exception()
+
+
+def is_turn_over(shot: System, constraints: ShotConstraints) -> bool:
+    assert constraints.cueable is not None
+    cue = constraints.cueable[0]
+
+    # Find when the second ball is first hit by the cue-ball
+
+    ball_hits = filter_events(
+        shot.events,
+        by_type(EventType.BALL_BALL),
+        by_ball(cue),
+    )
+
+    hits = set()
+    for event in ball_hits:
+        hits.add(_other(cue, event))
+        if len(hits) == 2:
+            break
+    else:
         return True
 
-    # Count rails that cue ball hits
-    cue_cushion_hits = filter_events(
+    # Now calculate all cue-ball cushion hits before that event
+
+    cushion_hits = filter_events(
         shot.events,
         by_type(EventType.BALL_LINEAR_CUSHION),
-        by_ball("cue"),
+        by_ball(cue),
+        by_time(event.time, after=False),
     )
 
-    # Count rails that object ball hits
-    object_cushion_hits = filter_events(
-        shot.events,
-        by_type(EventType.BALL_LINEAR_CUSHION),
-        by_ball("object"),
-    )
-
-    return len(cue_cushion_hits) + len(object_cushion_hits) != 3
+    return len(cushion_hits) < 3
 
 
 def is_game_over(
@@ -48,13 +65,23 @@ def is_game_over(
     return score[active.name] == win_condition
 
 
-class _SumToThree(Ruleset):
+def next_cue(current_cue: str, num_players: int) -> str:
+    assert current_cue in ("white", "yellow", "red")
+    assert num_players in (2, 3)
+
+    if num_players == 3:
+        return {"white": "yellow", "yellow": "red", "red": "white"}[current_cue]
+
+    return "white" if current_cue == "yellow" else "yellow"
+
+
+class _ThreeCushion(Ruleset):
     def __init__(self, *args, win_condition: int = 10, **kwargs):
         self.win_condition = win_condition
         Ruleset.__init__(self, *args, **kwargs)
 
     def build_shot_info(self, shot: System) -> ShotInfo:
-        turn_over = is_turn_over(shot)
+        turn_over = is_turn_over(shot, self.shot_constraints)
         score = self.get_score(self.score, turn_over)
         game_over = is_game_over(
             score,
@@ -77,13 +104,24 @@ class _SumToThree(Ruleset):
         return ShotConstraints(
             ball_in_hand=BallInHandOptions.NONE,
             movable=[],
-            cueable=["cue"],
-            hittable=("object",),
+            cueable=["white"],
+            hittable=tuple(),
             call_shot=False,
         )
 
     def next_shot_constraints(self, shot: System) -> ShotConstraints:
-        return self.shot_constraints
+        assert (cueable := self.shot_constraints.cueable) is not None
+
+        if self.shot_info.turn_over:
+            cueable = [next_cue(cueable[0], len(self.players))]
+
+        return ShotConstraints(
+            ball_in_hand=BallInHandOptions.NONE,
+            movable=[],
+            cueable=cueable,
+            hittable=tuple(),
+            call_shot=False,
+        )
 
     def get_score(self, score: Counter, turn_over: bool) -> Counter:
         if turn_over:
@@ -102,5 +140,5 @@ class _SumToThree(Ruleset):
         if self.shot_info.turn_over:
             self.log.add_msg(f"{self.last_player.name} is up!", sentiment="good")
 
-    def copy(self) -> _SumToThree:
+    def copy(self) -> _ThreeCushion:
         raise NotImplementedError("ThreeCushion copy needs to be implemented")
